get_gradient_points: take 'angle' as a double instead of a float
The originating values used to compute an angle were generally doubles and
computation was done in `get_gradient_points` as if this value was a double
(e.g. calling `sin` and `cos`). Treating it consistently as a double squashes a
number of -Wfloat-conversion compiler warnings and improves precision.

+++ lib/common/utils.c
@@ -1743,8 +1743,7 @@ void setEdgeType (graph_t* g, int dflt)
  * By default, this assumes a left-hand coordinate system (for svg); if RHS = 2 flag
  * is set, use standard coordinate system.
  */
-void get_gradient_points(pointf * A, pointf * G, int n, float angle, int flags)
-{
+void get_gradient_points(pointf *A, pointf *G, int n, double angle, int flags) {
     int i;
     double rx, ry;
     pointf min,max,center;
